pool: Randomize first mac allocation (#447)
* pool-manager/pool_test: Assert expected MACs from vm's interfaces

Currently the unit test assumes that the first MAC in the range is the
one that is going to be allocated.
In preparation to changing the algorithm of the MAC allocation to not
start with the start of the range, changing the check to expect the MACs
set on the VM's interfaces, and only them.
This keeps the allocations logic and makes sure that the pool indeed
keeps the MACs handed to the VM's interfaces.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pool: Change MAC allocation strategy to start from random assignment

Previously, the MAC allocation strategy relied on sequential assignment
and started from the beginning of the range.

This approach can cause issues in specific use cases where
multi-clusters are merged into one, increasing the likelihood of MAC
conflict .
Changing the algorithm to be choose the first MAC assignment at random,
then continue with sequential assignment. This to ensures more robust
MAC allocation, but also keeps assignment relatively predictable and
easy to debug.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pool-manager/pool_test: Increase pool range, reduce false positives

Since now MAC allocation is done randomly, it is better to increase the
MAC pool range to be large, to greatly reduce cases where the randomly
allocated MAC is the one accidentally expected by the tests.
Also change managedNamespaceMAC, unmanagedNamespaceMAC to be outside of
this range to reduce rare cases where the MAC allocator will assign that
MAC to the interfaces, creating a forbidden duplication.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

---------

Signed-off-by: Ram Lavi <ralavi@redhat.com>