tests/vms: Remove legacy configmap tests
the legacy configmap was removed in v1.23.0 [0],[1] and thus is no
longer supported. Now it is time to remove the code related to the
upgrade process from a version where the configmap exists to a version
where it is no longer supported. Doing so will simplify kubemapool code.

This commit removes the e2e test, as a primer to its code removal in
following commits.

[0]
https://github.com/k8snetworkplumbingwg/kubemacpool/releases/tag/v0.23.0
[1] https://github.com/k8snetworkplumbingwg/kubemacpool/pull/269

Signed-off-by: Ram Lavi <ralavi@redhat.com>