pool: Optimize InitMaps namespace filtering to reduce API calls
Currently for every ns we fetch the webhook config, which is a redundant
API call.
Refactor to only call the webhook once, then pass to the functions that
checks if the namespace is managed, using the private function
isNamespaceManagedFromObject.
This significantly increases the performance of the InitMaps function in
cases where many namespaces are managed.

The logic of the functions is not changed.

Signed-off-by: Ram Lavi <ralavi@redhat.com>