pool: Refactor IsNamespaceManaged
Refactoring IsNamespaceManaged function, using a new private function
isNamespaceManagedFromObject to check purely the "is namespace managed"
logic, but without kubeAPI calls, by getting everything it needs as
input
params.

This will be used in future commits in order to check the logic of
InitMaps without using kubeAPI on every call to
isNamespaceManagedFromObject.

Signed-off-by: Ram Lavi <ralavi@redhat.com>